pixel change int to short

+++ Graphics/Pixel.hpp
@@ -3,8 +3,8 @@
 
 struct Pixel
 {
-	unsigned int red;
-	unsigned int green;
-	unsigned int blue;
+	unsigned short red;
+	unsigned short green;
+	unsigned short blue;
 };
 #endif // !__PIXEL_HPP
